fix(highlight): split grammar registry and satisfy CI checks (#857)
